feat: 테마 삭제 요청시 해당 테마의 예약 존재할 경우 예외 처리 추가
- reservationRepository.findByThemeId(Long themeId) 추가 (구현체 작성)
- reservationService.getReservationByThemeId(Long themeId) 추가
- themeController.deleteTheme() 수정
- 관련 Repository, Service 테스트 작성

+++ src/main/java/nextstep/service/ThemeService.java
@@ -2,9 +2,13 @@ package nextstep.service;
 
 import nextstep.dto.ThemeRequest;
 import nextstep.exception.ThemeNotFoundException;
+import nextstep.exception.ThemeReservationExistsException;
+import nextstep.model.Reservation;
 import nextstep.model.Theme;
 import nextstep.web.JdbcTemplateThemeRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class ThemeService {
@@ -24,7 +28,10 @@ public class ThemeService {
                 .orElseThrow(() -> new ThemeNotFoundException(id));
     }
 
-    public void deleteTheme(Long id) {
+    public void deleteTheme(Long id, List<Reservation> reservations) { // 다른 서비스에서 엔티티 사용??
+        if (reservations.size() != 0) {
+            throw new ThemeReservationExistsException(id);
+        }
         themeRepository.deleteById(id);
     }
 }
